Extract series reconciliation into pipeline module
